inventory controller created

+++ src/main/java/com/ecom/demo/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.ecom.demo.service;
 
+import com.ecom.demo.entity.Inventory;
 import com.ecom.demo.entity.Products;
 import com.ecom.demo.repository.ProductsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +26,13 @@ public class ProductService {
 
 
     public void deleteProductById(int id){
-        productsRepository.deleteById(id);
+        Optional<Products> productOptional = productsRepository.findById(id);
+
+        if (productOptional.isPresent()) {
+            productsRepository.deleteById(id);
+        } else {
+            System.out.println("Product with ID " + id + " not found.");
+        }
     }
 
 
@@ -39,7 +46,6 @@ public class ProductService {
         if (productOptional.isPresent()) {
             productsRepository.save(products);
         } else {
-            // Handle the case where the product with the given ID does not exist
             System.out.println("Product with ID " + id + " not found.");
         }
     }
